+++ src/index.ts
@@ -4,6 +4,7 @@ import unformatMoney from './utils/unformatMoney';
 import textEllipsis from './utils/textEllipsis';
 import formatCep from './utils/formatCep';
 import capitalizeFirstLetter from './utils/capitalizeFirstLetter';
+import insert from './utils/insert';
 
 import isCnpj from './utils/isCnpj';
 import isCpf from './utils/isCpf';
@@ -16,4 +17,5 @@ export {
   isCnpj,
   isCpf,
   capitalizeFirstLetter,
+  insert,
 };